feat: reply count change immediately in UserLikePage
