+++ src/pages/UserLikesPage.jsx
@@ -12,14 +12,14 @@ import SideBar from "components/SideBar";
 import SwitchBar from "components/SwitchBar";
 import UserInfo from "components/UserInfo";
 import { UserTweetItem } from "components/TweetItem";
-import { getUserLikedTweets } from "api/userAuth";
+import { InfoContext } from "contexts/InfoContext";
 
 function UserLikesPage() {
-  const [userLikedTweets, setUserLikedTweets] = useState([]);
   const [currentPage, setCurrentPage] = useState("likes");
 
   const { isTweetModalShow, handleTweetClick, isReplyModalShow } =
     useContext(TweetContext);
+  const { getUserLikedTweetsAsync, userLikedTweets } = useContext(InfoContext);
 
   const navigate = useNavigate();
   const location = useLocation();
@@ -39,17 +39,9 @@ function UserLikesPage() {
 
   // useEffect
   useEffect(() => {
-    const getUserLikedTweetsAsync = async () => {
-      try {
-        const userLikedTweetsData = await getUserLikedTweets(pageUserId);
-        setUserLikedTweets(userLikedTweetsData);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
     getUserLikedTweetsAsync();
-  }, [pageUserId]);
+    //eslint-disable-next-line
+  }, []);
 
   return (
     <div className="d-flex">
